requirements.txt
fbchat
request
revChatGPT